chore: remove cached bytecode
